requirements.txt
# LLM & LangChain core
langchain==0.3.19
langchain-openai==0.3.9
langchain-community==0.3.18
langchain-core==0.3.51
openai==1.70.0

# RAG
faiss-cpu==1.9.0.post1
tiktoken==0.8.0
pypdf==4.3.1

# Agent web search tool
tavily-python==0.5.0

# LangGraph (multi-agent phase)
langgraph==0.3.29

# Frontend
streamlit==1.44.0

# Utilities
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.3
pyfarmhash
langchain-pinecone==0.2.0
pinecone-client==5.0.1